Restructured actions.
