--- common/actions/history.js
@@ -1,8 +1,0 @@
-import * as types from '../constants/ActionTypes';
-
-export function reAddSong(index) {
-  return {
-    type: types.ADD_SONG_FROM_HISTORY,
-    index
-  };
-}
