+++ common/actions/historyActions.js
@@ -1,0 +1,3 @@
+// import * as types from '../constants/ActionTypes';
+
+// actions for history reducer will go here
