--- common/actions/nextSong.js
@@ -1,7 +1,0 @@
-import * as types from '../constants/ActionTypes';
-
-export default function nextSong() {
-  return {
-    type: types.NEXT_SONG
-  };
-}
