+++ common/actions/queueActions.js
@@ -12,3 +12,15 @@ export function nextReady() {
     type: types.NEXT_READY
   };
 }
+
+export function playSong() {
+  return {
+    type: types.PLAY_SONG
+  };
+}
+
+export function pauseSong() {
+  return {
+    type: types.PAUSE_SONG
+  };
+}
