--- common/actions/song.js
@@ -1,8 +1,0 @@
-import * as types from '../constants/ActionTypes';
-
-export function upvoteSong(index) {
-  return {
-    type: types.UPVOTE_SONG,
-    index
-  };
-}
